Push main content into index.js, propagate function name changes.
